requirements.txt
biopython=1.73=py37h7b6447c_0
matplotlib=3.0.3=py37h5429711_0
seaborn=0.9.0=py37_0